% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mETHYLotest.EPIC.DMP.report.R
\name{mETHYLotest.EPIC.DMP.report}
\alias{mETHYLotest.EPIC.DMP.report}
\title{Generate HTML Reports for DMP Signatures}
\usage{
mETHYLotest.EPIC.DMP.report(
  dmp_results,
  beta,
  pheno,
  output_directory = file.path(getwd(), "EpicPipeline", "DMP_Reports"),
  signatures_folder = NULL,
  signature_comparison_source = NULL,
  adjPVal = 0.05,
  arraytype = "EPICv1",
  top_n_heatmap = 50,
  top_n_windows = 20,
  rmd_template_path = NULL
)
}
\arguments{
\item{dmp_results}{List of dataframes (ChAMP output).}

\item{beta}{Numeric matrix of Beta values.}

\item{pheno}{Vector of phenotype labels.}

\item{output_directory}{Path to save HTML files.}

\item{signatures_folder}{(Optional) Path to folder containing .txt signature files.}

\item{signature_comparison_source}{(Optional) Name of the comparison in 'dmp_results' to use for statistics when in Signature Mode.}

\item{adjPVal}{Threshold for standard mode.}

\item{arraytype}{"EPICv1", "EPICv2", or "450K".}

\item{rmd_template_path}{Path to the Rmd file (optional override).}
}
\description{
Generates HTML reports. Can works in two modes:
1. Standard Mode: Iterates over dmp_results, filters by adjPVal, creates reports.
2. Signature Mode: If 'signatures_folder' is provided, generates reports for each .txt signature file found.
}
