% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mETHYLotest.NGS.utils.R
\name{mETHYLotest.NGS.ApplyOffsetToObj}
\alias{mETHYLotest.NGS.ApplyOffsetToObj}
\title{Apply coordinate offset to methylRawList}
\usage{
mETHYLotest.NGS.ApplyOffsetToObj(myobj, offset = 1L)
}
\arguments{
\item{myobj}{A \code{methylRawList} object from \code{methylKit::methRead}.}

\item{offset}{Integer. Value to add to start and end (default: 1).}
}
\value{
A corrected \code{methylRawList} object.
}
\description{
Adds a fixed offset to start and end coordinates of all samples
in a methylRawList object. Used to convert 0-based BED coordinates
to 1-based coordinates expected by methylKit.
}
