% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mETHYLotest.NGS.DiffMeth.UI.R
\name{mETHYLotest.NGS.DiffMeth.UI}
\alias{mETHYLotest.NGS.DiffMeth.UI}
\title{Interactive Differential Methylation Results UI}
\usage{
mETHYLotest.NGS.DiffMeth.UI(diff_list, output_dir)
}
\arguments{
\item{diff_list}{Named list of methylDiff objects.}

\item{output_dir}{Directory to save filtered results.}
}
\value{
NULL (invisibly). Blocks until closed.
}
\description{
Interactive Differential Methylation Results UI
}
