% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mETHYLotest.NGS.QC.GenerateReport.R
\name{mETHYLotest.NGS.GenerateQCReport}
\alias{mETHYLotest.NGS.GenerateQCReport}
\title{Generate a static HTML QC report}
\usage{
mETHYLotest.NGS.GenerateQCReport(
  df_meth,
  df_controls,
  df_controls_all = NULL,
  df_stats,
  df_global_meth = NULL,
  min_cov,
  excluded_samples,
  project_dir = NULL,
  output_file
)
}
\arguments{
\item{df_meth}{Dataframe of methylation by chromosome}

\item{df_controls}{Dataframe of control genome stats (current iteration)}

\item{df_controls_all}{Dataframe of control genome stats for ALL samples (including excluded)}

\item{df_stats}{Dataframe of general position stats}

\item{df_global_meth}{Dataframe of global methylation per sample}

\item{min_cov}{The coverage threshold used}

\item{excluded_samples}{Vector of excluded sample IDs}

\item{project_dir}{Project directory (for loading saved RDS files in the report)}

\item{output_file}{Path to the output HTML file}
}
\description{
Generate a static HTML QC report
}
