% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mETHYLotest.NGS.utils.R
\name{mETHYLotest.NGS.get_C_counts}
\alias{mETHYLotest.NGS.get_C_counts}
\title{Calculate, plot and export total C vs T counts per sample}
\usage{
mETHYLotest.NGS.get_C_counts(
  methyl_obj,
  output_dir = NULL,
  plot_name = "total_C_counts.png",
  excel_name = "total_C_counts.xlsx",
  scale_y_log10 = FALSE
)
}
\arguments{
\item{methyl_obj}{A `methylRawList` object.}

\item{output_dir}{Directory path to save outputs. If NULL, nothing is saved.}

\item{plot_name}{Name of the output plot file (e.g., "total_C_counts.png").}

\item{excel_name}{Name of the output Excel file (e.g., "total_C_counts.xlsx").
If NULL, no Excel file is generated.}

\item{scale_y_log10}{Logical. If TRUE, uses a log10 scale for the plot.}
}
\value{
A data.frame (Long format) used for plotting.
}
\description{
Aggregates the total number of Methylated Cytosines (numCs) and
Unmethylated Cytosines (numTs) for each sample.
Generates a stacked bar plot and an Excel report with percentages.
}
