% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mETHYLotest.NGS.ProjectUI.R
\name{mETHYLotest.NGS.ProjectUI}
\alias{mETHYLotest.NGS.ProjectUI}
\title{Launch mETHYLotest Project Setup UI (WGBS & Long Read)}
\usage{
mETHYLotest.NGS.ProjectUI(prefill_pheno = NULL)
}
\value{
Absolute path of the project directory (invisibly).
}
\description{
Launches a Shiny Dashboard interface to initialize or reload a mETHYLotest
project for WGBS and Long Read methylation analysis.
}
