% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mETHYLotest.NGS.QC.R
\name{mETHYLotest.NGS.QC}
\alias{mETHYLotest.NGS.QC}
\title{Quality Control and Descriptive Analysis for Methylation Data}
\usage{
mETHYLotest.NGS.QC(
  methyl_obj,
  output_base_dir,
  chromosomes = c(paste0("chr", 1:22), "chrX", "chrY", "chrM"),
  current_min_cov = 1,
  unite_destrand = FALSE,
  save_summary = TRUE
)
}
\arguments{
\item{methyl_obj}{A methylRawList object (filtered or not)}

\item{output_base_dir}{The base directory for figures}

\item{chromosomes}{Vector of chromosomes to analyze}

\item{current_min_cov}{Integer, the coverage threshold used (for logging)}

\item{save_summary}{Boolean. If TRUE, saves all stats tables to an Excel file.}
}
\value{
A list containing the dataframes used for the QC UI (df_meth, df_controls, df_stats)
}
\description{
Performs a comprehensive QC on a methylRawList object. It generates plots and
aggregates statistics, optionally saving all metrics into a summary Excel file.
}
