% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mETHYLotest.utils.R
\name{mETHYLotest.utils.HarmonizeEpicV2toV1}
\alias{mETHYLotest.utils.HarmonizeEpicV2toV1}
\title{Harmonize EPICv1 and EPICv2 (compatibility wrapper)}
\usage{
mETHYLotest.utils.HarmonizeEpicV2toV1(
  myLoad_v1,
  myLoad_v2,
  duplicate_strategy = c("mean", "first")
)
}
\arguments{
\item{myLoad_v1}{ChAMP object for EPICv1 samples.}

\item{myLoad_v2}{ChAMP object for EPICv2 samples.}

\item{duplicate_strategy}{Passed to
\code{\link{mETHYLotest.utils.HarmonizeArrays}}.}
}
\description{
Wrapper around \code{\link{mETHYLotest.utils.HarmonizeArrays}} for the
EPICv1 + EPICv2 two-array case. Kept for backward compatibility.
}
\seealso{
\code{\link{mETHYLotest.utils.HarmonizeArrays}}
}
